+++ runtime/src/lib.rs
@@ -60,7 +60,7 @@ pub use pallet_midnight::{TransactionTypeV2, pallet::Call as MidnightCall};
 pub use pallet_midnight_system::Call as MidnightSystemCall;
 pub use pallet_session_validator_management::{self, Config};
 use pallet_session_validator_management::{
-	CommitteeInfo, CurrentCommittee, migrations::authority_keys::UpgradeCommitteeMember,
+	CommitteeInfo, CurrentCommittee, migrations::v2::UpgradeCommitteeMember,
 };
 pub use pallet_timestamp::Call as TimestampCall;
 pub use pallet_version::VERSION_ID;
@@ -2462,7 +2462,7 @@ mod tests {
 		};
 		use authority_selection_inherents::CommitteeMember;
 		use frame_support::BoundedVec;
-		use pallet_session_validator_management::migrations::authority_keys::UpgradeCommitteeMember;
+		use pallet_session_validator_management::migrations::v2::UpgradeCommitteeMember;
 		use parity_scale_codec::Encode;
 		use sidechain_domain::ScEpochNumber;
 		use sp_core::Pair;